Moving optimum implementation, several lineagesfor simulations

+++ simulator/src/lib/random.hpp
@@ -3,7 +3,7 @@
 #include <random>
 
 // Random generator engine with seed 0.
-double seed{0};
+u_long seed{0};
 std::default_random_engine generator(seed);
 
 std::normal_distribution<double> normal_distrib(0.0, 1.0);
